Fix Bugs In Route Planning

+++ Lines/Lines.cpp
@@ -22,15 +22,15 @@
 #include "Lines.h"
 
 //多源最短路
-std::vector<std::vector<double>>dfs_map;
+std::vector<std::vector<float>>dfs_map;
 
 //返回最小值
-inline double Min(double a, double b) {
+inline float Min(float a, float b) {
 	return a > b ? b : a;
 }
 
 //由已有有向图生成多源最短路,时间复杂度O(n³)
-std::vector<std::vector<double>>Floyd(std::vector<std::vector<double>>map) {
+std::vector<std::vector<float>>Floyd(std::vector<std::vector<float>>map) {
 	int n = map[0].size();
 	for (int k = 0; k < n; k++) {
 		for (int i = 0; i < n; i++) {
@@ -62,14 +62,14 @@ Dfs_Return Dfs(int start, int from, std::vector<int>to, std::vector<bool>visit, 
 			}
 		}
 	}
-	double Min_Distance = INF;
+	float Min_Distance = INF;
 	Dfs_Return tmp;
 	for (int i = 0; i < n; i++) {
 		if (!visit[i]) {
 			visit[i] = true;
 			tmp = Dfs(start, to[i], to, visit, step + 1);
 			visit[i] = false;
-			double New_Distance = Min(tmp.cost + dfs_map[from][to[i]], INF);
+			float New_Distance = Min(tmp.cost + dfs_map[from][to[i]], INF);
 			if (New_Distance < Min_Distance) {
 				Min_Distance = New_Distance;
 				ans = tmp;
@@ -82,7 +82,7 @@ Dfs_Return Dfs(int start, int from, std::vector<int>to, std::vector<bool>visit, 
 }
 
 //返回到各目的地的先后
-std::vector<int>Get_Primary_Order(int from, std::vector<int>to, std::vector<std::vector<double>>map) {
+std::vector<int>Get_Primary_Order(int from, std::vector<int>to, std::vector<std::vector<float>>map) {
 	std::vector<int>ans;
 	int n = map[0].size();
 	dfs_map = Floyd(map);
@@ -99,10 +99,10 @@ std::vector<int>Get_Primary_Order(int from, std::vector<int>to, std::vector<std:
 }
 
 //堆优化的Dijkstra,但是仅用于求路径顺序,返回反向的不包含起点但包含终点的路径顺序
-std::vector<int>Dijkstra(int from, int to, std::vector<std::vector<double>>map) {
+std::vector<int>Dijkstra(int from, int to, std::vector<std::vector<float>>map) {
 	//初始化
 	std::vector<int>ans;
-	std::vector<double>from_point;
+	std::vector<float>from_point;
 	std::priority_queue<Distance_To_Start>queue;
 	int n = map[0].size();
 	Distance_To_Start head{ from,0,-1 };
@@ -134,7 +134,7 @@ std::vector<int>Dijkstra(int from, int to, std::vector<std::vector<double>>map) 
 }
 
 //输出具体路径信息
-std::vector<int>Get_Order(int from, std::vector<int>to, std::vector<std::vector<double>>map) {
+std::vector<int>Get_Order(int from, std::vector<int>to, std::vector<std::vector<float>>map) {
 	std::vector<int>primary_order = Get_Primary_Order(from, to, map);
 	std::vector<int>tmp;
 	std::vector<int>ans;
